模拟线上的测评逻辑

+++ src/main/java/com/huawei/hwcloud/gaussdb/data/store/race/vo/DeltaPacket.java
@@ -18,7 +18,7 @@ public class DeltaPacket implements Serializable {
 
     private long version;
 
-    private long deltaCount;
+    private short deltaCount;
 
     private List<DeltaItem> deltaItem;
